Removing duplicate version of EventCountdown.
